convert azimuth to float

+++ main.cpp
@@ -123,19 +123,19 @@ void reverse32(uint32_t &byte){
 
 void printPacket( Packet &packet){
     reverse16(packet.block0.azimuth);
-    std::cout<<packet.block0.azimuth/100<< "\n";
+    std::cout<<static_cast<float>(packet.block0.azimuth)/100<< "\n";
     reverse16(packet.block1.azimuth);
-    std::cout<<packet.block1.azimuth/100<< "\n";
+    std::cout<<static_cast<float>(packet.block1.azimuth)/100<< "\n";
     reverse16(packet.block2.azimuth);
-    std::cout<<packet.block2.azimuth/100<< "\n";
+    std::cout<<static_cast<float>(packet.block2.azimuth)/100<< "\n";
     reverse16(packet.block3.azimuth);
-    std::cout<<packet.block3.azimuth/100<< "\n";
+    std::cout<<static_cast<float>(packet.block3.azimuth)/100<< "\n";
     reverse16(packet.block4.azimuth);
-    std::cout<<packet.block4.azimuth/100<< "\n";
+    std::cout<<static_cast<float>(packet.block4.azimuth)/100<< "\n";
     reverse16(packet.block5.azimuth);
-    std::cout<<packet.block5.azimuth/100<< "\n";
+    std::cout<<static_cast<float>(packet.block5.azimuth)/100<< "\n";
     reverse16(packet.block6.azimuth);
-    std::cout<<packet.block6.azimuth/100<< "\n";
+    std::cout<<static_cast<float>(packet.block6.azimuth)/100<< "\n";
     std::cout<<static_cast<int>(packet.block0.flagEE);
     std::cout<<static_cast<int>(packet.block0.flagFF);
     std::cout<<static_cast<int>(packet.block1.flagEE);
